куча всего

+++ internal/controller/constants.go
@@ -7,7 +7,7 @@ const (
 	ColorGreen = "positive"
 )
 
-var AccessChatIds = []int64{2000000002, 2000000001}
+var AccessChatIds = []int{2000000002, 2000000001}
 
 const (
 	RoleAdmin      = "admin"
